examples/requirements.txt
pandas==1.*
oceania-query-fasta>=0.1.6